Avoid UB on invalid borrows
As the mere existence of aliasing mutable references, not their use
is/may be (the aliasing model is not fully defined yet) an UB, make sure
the references are created only after the check for validity is done and
any illegal situations are prevented by a panic.

It's unlikely the compiler is able to take advantage of in practice,
though.